backend/requirements.txt

fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
openai==1.3.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
